agent-config: rename agent to aa - rename agent to aa to reflect the real config
